refact : Coordinate 위도 경도를 Int 값으로 관리하도록 변경
위도, 경도 값은 소숫점 6자리까지 표현되고 소숫점을 모두 정수로 만들어줄 경우에 Int 범위를 초과하지 않기 때문에 Int로 변경하여 데이터를 관리하도록 변경
러닝 데이터는 매일 생성되고, 위도, 경도는 리스트의 값으로 서버에 무수히 많게 저장될 예정이기 때문에 Int를 통해 비용을 낮추기로 결정

+++ RunWithMe/domain/src/main/java/com/side/domain/model/Coordinate.kt
@@ -4,6 +4,6 @@ import java.io.Serializable
 
 
 data class Coordinate(
-    val latitude: Double,
-    val longitude: Double
+    val latitude: Int,
+    val longitude: Int
 )
